DHHS Collaboration Network

This data set contains the DHHS Collaboration network. This is an undirected network with 54 nodes and 477 edges.

Format

A network object with 54 nodes and 477 edges

Details

These data were obtained from an evaluation of tobacco control leadership in the Department of Health and Human Services in 2005. The network data represent collaboration ties among 54 tobacco control leaders working in eleven different agencies.

Node characteristics:

vertex.names - ID with agency code and member number

agency - Agency code:

0 - ACF

1 - AHRQ

2 - CDC

3 - CMS

4 - FDA

5 - HRSA

6 - IHS

7 - NIH

8 - OGC

9 - OS

10 - SAMHSA

Edge characteristics:

collab - Level of collaboration (only in TCcoll):

0 - No collaboration

1 - Share info only

2 - Collaborate informally

3 - Collaborate formally

4 - Collaborate formally on multipel projects

References

Leischow, S. J., Luke, D.A., et al. (2010). Mapping U.S. government tobacco control leadership: Networked for success? Nicotine & Tobacco Research, 12, 888-894. Harris, J. K., et al. (2012). Interpersonal influence among public health leaders in the United States Department of Health and Human Services. Journal of Public Health Research, 1, 67-74.