9-11 hijackers network

This network shows the contacts among the terrorists who carried out the 9-11 hijackings in 2001.

Format

A network object with 19 vertices and 27 edges

Details

This network data set was first constructed and analyzed by Valdis Krebs. Using publically available data at the time, he constructed this contact network to illustrate the contact ties that the 19 hijackers had prior to the attack.

Node characteristics:

name - Full name

flight - Flight number

References

Krebs, Valdis. Uncloaking Terrorist Networks. First Monday, [S.l.], apr. 2002. ISSN 13960466. Available at: http://firstmonday.org/ojs/index.php/fm/article/view/941/863.