State tobacco control organizational networks

This set of networks contains organizational network data collected from the Indiana state tobacco control program in 2010. Each network is made up of the same 25 organizational nodes. Three of the networks contain interorganizational relationship data: contact, collaboration, and dissemination.

Format

A list that contains four network (statnet) objects

Details

These data were obtained as part of an evaluation of state tobacco control programs funded by the Centers of Disease Control and Prevention (CDC). Three different organizational relationships were measured and are stored in network objects within the TCnetworks list object. TCcnt contains the contact tie. A tie exists when the two organizations had contact with each other. TCcoll contains the collaboration tie. A tie exists when the two organizations formally collaborated with each other. Finally, TCdiss contains the dissemination tie. A tie exists when two organizations talked to each other about the CDC evidence-based guideline Best Practices for Tobacco Control. The TCdist network stores the physical distance (in square root of the miles) between each pair of organizations in the Indiana tobacco control program.

Node characteristics:

vertex.names - Organization name

agency_cat - Type of agency:

1 - Lead agency

2 - Contractor

3 - Coalition

4 - Advocacy

5 - Government

6 - National partner

agency_lvl - Level of agency:

1 - Local

2 - State

3 - National

lead_agency - Lead agency (0=No, 1=Yes)

tob_yrs - Number of years doing tobacco control

Edge characteristics:

contact - Frequency of contact (only in TCcnt):

1 - Yearly

2 - Quarterly

3 - Monthly

4 - Weekly

5 - Daily

collaborate - Level of collaboration (only in TCcoll):

0 - No collaboration

1 - Share info only

2 - Collaborate informally

3 - Collaborate formally

4 - Collaborate formally on multipel projects

distance - Distance between agencies, in miles (only in TCdist)

References

Luke, D.A., Wald, L. M., Carothers, B. J., Bach, L. E., Harris, J. K. (2013). Network influences on dissemination of evidence-based guidelines in state tobacco control programs. Health Education & Behavior, 40, 33S-42S.